Changes to bring the Razor CLI and RESTful API in line with the new-style RESTful API that was recently agreed to.  Includes changes to limit the use of "remove all" from the RESTful API (but leave that functionality in place from the CLI), the use of flags to pass in additional arguments/options to the "get" operator for the BMC and Node slices, removal of the UUID argument from the "bmc register" request (a UUID is now autogenerated based on the BMC's MAC address), and a restructuring of the "tag" and "tag matcher" relationship (from a URI perspective).  Also made changes to restrict the use of "aliases" from REST/HTTP (they are no longer supported/allowed) and to restrict access to some slice-related functionality that was never intended to be accessed via REST/HTTP (the "/config/dbcheck" resource, for example).  The changes made to the URI structure for "tag matchers" also required a change to the constructor for the underlying "ProjectRazor::Tagging::TagMatcher" object (and to the corresponding lines that used this constructor in the "ProjectRazor::Tagging::TagRule" object.  Finally, made changes to the rspec tests to fix issues caused by all of these changes (mostly having to do with the new "tag matcher" URI pattern).

+++ lib/project_razor/slice/active_model.rb
@@ -55,6 +55,7 @@ module ProjectRazor
 
       def get_active_model_logs
         @command = :get_active_model_logs
+        raise ProjectRazor::Error::Slice::MethodNotAllowed, "Cannot view Active Model logs via REST" if @web_command
         # the UUID is the first element of the @command_array
         uuid = @prev_args.peek(1)
         active_model = get_object("active_model_instance", :active, uuid)
@@ -64,6 +65,7 @@ module ProjectRazor
       end
 
       def remove_all_active_models
+        raise ProjectRazor::Error::Slice::MethodNotAllowed, "Cannot remove all Active Models via REST" if @web_command
         raise ProjectRazor::Error::Slice::CouldNotRemove, "Could not remove all Active Models" unless get_data.delete_all_objects(:active)
         slice_success("All active models removed", :success_type => :removed)
       end
@@ -79,6 +81,8 @@ module ProjectRazor
       end
 
       def get_logview
+        @command = :get_logview
+        raise ProjectRazor::Error::Slice::MethodNotAllowed, "Cannot view Active Model logs via REST" if @web_command
         active_models = get_object("active_models", :active)
         log_items = []
         active_models.each { |bp| log_items = log_items | bp.print_log_all }
